Adding experimental support for alcazar.
This is switched on by adding "alcazar: True" to the config file. It enables
body text extraction with the alcazar library (install it with pip).

Currently, the html passing through the pipeline into the lett file
may not be done correctly with alcazar, and I don't know if it affects
the bitextor document aligner. If you don't use alcazar, everything should
be as before.