Fixes and changes in dic-based doc and hunalign
Fix dic-based docalign when number of shards were not equal (necessary
files which were being expected to exist in 'all' rule were not being
correctly generated/managed). Added threshold to hunalign normalized
in order to be compatible with bleualign range: [0.0,1.0]. ExternalMT
docalign working with hunalign, but dic-based docalign with bleualign
not allowed.