Fix CMakeLists.txt CMAKE_CXX_FLAGS
When a previous value is set in CMAKE_CXX_FLAGS, maybe is necessary in order to proceed with the build. This cannot be done if the previous value is not used and is directly set to a constant value. This fix solves different problems (e.g. ICU headers not placed in default system directories, but in a custom path)