linguacrawl integration, conda build and fixes
The crawler linguacrawl has been integrated, and those options which
were only creepy-specific, now are shared (new configuration options
has been added as well). Due to the fact that linguacrawl is far
different from the others crawlers, specific logic has been needed
in order to be correctly integrated. Conda build scripts ready to
generate Bitextor builds (Bitextor 7, which is in master right now,
and 8, which is the current branch). Those scripts allows to
automatizate completely the generation of a build for Bitextor,
resulting in a file which might be uploaded (is the idea) to the
official repositories of Anaconda. These scripts attempt to handle
some known problems, but even so, problems might happen due to the
fact that is a first version. Different fixes have been applied:
relative call to ~/go/bin has been replaced with a relative call to
the binaries themselves, but in the case of not find the relative
binaries, the initial call will be made; in those cases when no
common shards were generated, neither segalign or docalign were
executed in any of the possible configurations, and this only happens
in the case of very different URLs and bad luck because no shard id
was equal in any of the provided languages, and in order to fix this
situation, a fake shard is generated only if is necessary, and after
the generation of this shard, the rules will be updated and lead to
a correct execution of the pipeline, avoiding to break; Biroamer
information was being discarted, which initially was intentional, but
it is not the best due to the fact that in some cases fails, but
finish correctly, and we could have though that it finished correctly
when did not