Fixes and args constraint
Deferred and Biroamer cannot be enabled at the same time, and so
has been configured as a constraint in the arguments processing.
Tests have been fixed according to this constraint. Conda build
post-link has been modified in order to create a soft link of
Bitextor since is necessary to know the location when a
configuration file is being created or modified