Multiple fixes
FTFY deprecated parameter was causing the preprocessing to break
and was not possible to notice it since there was no error nor
message.

Sharding processing for empty or near-empty WARCs was not
working as expected when 1 thread was configured. In this case
what was happening was an infinite loop due to the only thread
was waiting for another thread to reach the shard checkpoint.

Tests are now checking if the no. lines is 0 since it is not
what we expect, and in that case, the test fails (before was
not failing)