Fix wget and structure distance scripts
Wget script was displaying the output of a command that is not necessary
to be displayed.

The structure distance script, which is executed when aligning documents
with the dictionary pipeline, was using a set as data structure and
iterating over it, and since sets does not preserve order, different
executions were returning different results, causing non-deterministic
behaviour.