Revert conda build to 'monolithic' package
Conda is not intended to work properly with other package managers
(e.g. pip). There are multiple problems when pip packages try to
work properly on a conda environment and they belong to different
packages, e.g. when different packages from conda install the same
pip package, the constraints are not kept, so there might be
conflicts.

The proper way to work with conda packages is to define all the
dependencies in meta.yaml, even pip dependencies. This might lead
to a harder manteinance, but if not, for complex packages might
be really hard to keep all the dependencies in a single package
manager like pip and create a conda package.