Split parallel config and warning about mgiza
parallelWorkers was being used in order to set the number of jobs
and threads of the tools at once. This behaviour has been split
and now parallelWorkers will set the threads of the tools and
parallelJobs will set the number of parallel jobs to execute.

When mgiza is executed in parallel, it does not generate a
deterministic output. This has been fixed using just 1 thread
for mgiza and warning the user about the non-deterministic output
from Bitextor if a greater value than 1 is used.