Automatic GPU allocation for jobs like NDA
Automatic GPU allocation is achieved through mutex mechanisms.
Once a job acquires the mutex, it allocates a device. Affected
jobs are NDA, vecalign and bicleaner.

Since the mutex mechanisms are done through PersistentDict
from pytools (recommended from the snakemake FAQ:
https://snakemake.readthedocs.io/en/stable/project_info/
faq.html#i-want-to-pass-variables-between-rules-is-that
-possible), the GPU allocation should work even if
multiple snakemake instances are running. This may be
useful since it allows to handle automatically the GPU
allocation for different configurations of Bitextor.

Other changes:
 - Typo in docs
 - Use of snakemake handlers (e.g. onstart)