First changes for refactoring paragraphs id
Instead of have the para id in the same file where are the
sentences, a new file is intended to be created and processed by
the segalign.
This is what is being done right now, but moving all the content
in only one file, what leads to extra processing and makes many
rules dependant on this feature.